Fix bad assumption in x86_expr_checkea_getregusage() that was causing bad code, internal errors, and mis-optimization as seen in Bugzilla Bug 13. Reported by: meor@softhome.net
Fix a few minor bugs (not heavily tested) due to use of opersize == 32 as a
comparison; this fails when opersize == 64 due to bits = 64, but the default
size is still 4 bytes, not 2.

There's still a couple cases here where we might end up with internal errors
due to a lack of variety in return values; I'll fix those in the next commit.

svn path=/trunk/yasm/; revision=1012