Add a final SO for any .text section. This is used to identify the end of the final function.
FIXME: This creates a symbol (label .n_so) which currently has to be
added to the symtab. Peter thinks it should work fine outside the table,
so perhaps this is a bug in elf. Currently without being in the table
instead of identifying the end of the function, it identifies offset 0.

svn path=/trunk/yasm/; revision=1146