Phase 1 of a 2 phase cleanup to elf's internal structure.
Historically elf.c had a lot of switches for which machines it supports.
This replaces it with a structure of information and function pointers
which is selected in much the same way the switch value was before. Then
all such switches are replaced with calls through the current machine
handler structure.

Phase 2 will relocate each machine's support to its own file.

svn path=/trunk/yasm/; revision=1155