Add work around for static-build libtool binaries to find preloaded modules without the presence of the associated ".la" file.  This caused the Win32 nightly snapshots to be unusable (as they're static-built libtool).
At present, a search for a preloaded module occurs first, before the ".la"
is even looked for.

Noticed by: Oliver <mybox@ollysoft.de>

* libtool.m4: Define new config.h variable LIBTOOL_LIBEXT that is the
extension used for static libraries (e.g. ".a").  This is the extension used
for preloaded symbols.

* ltdl.c (lt_dlopenext): Add first search function to find filename.a prior
to other searches (.la and .so), so preloaded symbols are hit first.

svn path=/trunk/yasm/; revision=1157