Generate DWARF2 information from asm source, closing #43.  Asm-level source debug information is generated if no file/loc directives are used.  Also will generate basic DWARF2 info/abbrev/aranges if not specified in source.  What's not handled is multiple code sections; I need to figure out how DWARF2 expects these to be generated.
The implementation of this refactors all the line generation into
dwarf2-line.c and makes dwarf2-dbgfmt.c contain only the core DWARF2
functions.

One challenge yet to be taken care of is how to test the automatic generation,
as the current working directory is saved into the output object file.

* dwarf64_2loc: Update so built-in info generation doesn't happen.

svn path=/trunk/yasm/; revision=1376