Follow-up to #80 and #81: "xchg ax, ax" can use the short 0x90 form as 16-bit operations don't clear the high bits of the 64-bit register, so this is effectively a NOP as well.
While I'm still looking at this, make "xchg rax, rax" not output a REX
prefix.

I thought about being extra-clever and have "xchg ax, ax" not output an
operand-size prefix, but thought better of it: if the user says "ax"
they're probably going to expect the prefix to be there.  Left in as
a comment for future readers of the code.

svn path=/trunk/yasm/; revision=1639