Fix #84: I broke the idiom of concating a decimal number to a register in the preproc (e.g mm%$x -> mm0) when I implemented a minor optimization (outputing values in hex to avoid the 2x decimal conversion overhead) in [1488].  Revert that optimization.
I believe the right way to do this is to add the GAS (and MASM) syntax
sugar of mm(0) etc. to NASM syntax, then any number (not just decimal)
can be used as the new idiom for doing this.  But that wouldn't unbreak
programs that already exist, so we're left with this state of affairs.
Sigh.

svn path=/trunk/yasm/; revision=1657