Push down GAS .ident directive handling into objfmts.  Note this makes it such that only ELF and COFF actually handle .ident; all other object formats silently ignore it.
In Win32/Win64, the .ident contents are dumped into the .rdata$zzz section
rather than the .comment section as in COFF and ELF; this is due to the
fact the GNU linker for Win32 is broken and puts .comment in the output
executable as a separate section, breaking the executable.  The .rdata$zzz
name matches that used by GAS to work around this same linker bug.

Reported and tracked down by: Peter Tanski <p.tanski@gmail.com>

svn path=/trunk/yasm/; revision=1712