Gas parser: Move instruction/prefix lookup from tokenizer to parser. Use the single token of lookahead to detect the label case.
This is significantly cleaner as it removes the special-casing of labels
in the tokenizer (so there is just a single identifier rule) and removes
the INSTDIR parser state (as this was only used to prevent instruction
lookup within other locations).

Also, ID and LABEL now provide the string length to the parser.  We needed
to do this for ID due to parse_check_insnprefix() needing the length, so
both were folded in for consistency.

svn path=/trunk/yasm/; revision=2164